+++ include/roynumber.h
@@ -26,7 +26,7 @@ double roy_parse_double(const char * str);
 
 // Converts the integer 'number' into a 'width'-width, 'base' character representation in the string 'dest'.
 // (The behavior is undefined if 'dest' or 'width' is insufficient, or 'base' is larger than 36 ([0-9a-z]).)
-char * roy_llong_to_string(char * dest, long long number, size_t base, size_t width, bool fill_zero);
+char * roy_llong_to_string(char * dest, int64_t number, size_t base, size_t width, bool fill_zero);
 
 // Converts the unsigned integer 'number' into a 'width'-width, 'base' character representation in the string 'dest'.
 // (The behavior is undefined if 'dest' or 'width' is insufficient, or 'base' is larger than 36 ([0-9a-z]).)
@@ -34,17 +34,17 @@ char * roy_ullong_to_string(char * dest, uint64_t number, size_t base, size_t wi
 
 // Returns 'dest' with the 'count' bits that begin at position 'position' set to the rightmost 'count' bits of 'src', 
 // the other bits are left unchanged.
-uint64_t roy_ullong_set_bits(uint64_t * dest, int position, size_t count, uint64_t src);
+uint64_t roy_ullong_set_bits(uint64_t * dest, size_t position, size_t count, uint64_t src);
 
 // Returns 'dest' with the 'count' bits that begin at position 'position' inverted 
 // (i.e., 1 changed into 0 and vice versa), the other bits are left unchanged.
-uint64_t roy_ullong_invert(long long * number, int position, size_t count);
+uint64_t roy_ullong_invert(uint64_t * number, size_t position, size_t count);
 
 // Returns the value of the 'width'-width unsigned integer 'number' rotated to the right by 'steps' bit positions.
-uint64_t roy_ullong_rotate_right(uint64_t * number, int steps, size_t width); 
+uint64_t roy_ullong_rotate_right(uint64_t * number, size_t steps, size_t width);
 
 // Returns the value of the 'width'-width unsigned integer 'number' rotated to the left by 'steps' bit positions.
-uint64_t roy_ullong_rotate_left(uint64_t * number, int steps, size_t width); 
+uint64_t roy_ullong_rotate_left(uint64_t * number, size_t steps, size_t width);
 
 // Counts '1' bits in unsigned integer 'number'.
 size_t roy_ullong_count_bit(uint64_t number);
